fix: support widened upload listing ids and signed restart timestamps
Keep UploadableListing aligned with API 15 by widening listing ids to ulong/u64 and accepting signed restart timestamps without losing legacy Mongo compatibility.

Preserve legacy v1 numeric ids, keep v2 string ids, and tighten upload validation/schema so new restart values stay within the signed i32 range while wide-id query paths avoid internal BSON mismatches.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ csharp/RemotePartyFinder/UploadableListing.cs
@@ -10,7 +10,7 @@ namespace RemotePartyFinder;
 [Serializable]
 [JsonObject(NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
 internal class UploadableListing {
-    public uint Id { get; }
+    public ulong Id { get; }
     public uint ContentIdLower { get; } // to retain backwards compatibility with old listings (stats), we stick to the lower bits
     public byte[] Name { get; }
     public byte[] Description { get; }
@@ -25,7 +25,7 @@ internal class UploadableListing {
     public ushort MinItemLevel { get; }
     public byte NumParties { get; }
     public byte SlotsAvailable { get; }
-    public uint LastServerRestart { get; }
+    public int LastServerRestart { get; }
     public ObjectiveFlags Objective { get; }
     public ConditionFlags Conditions { get; }
     public DutyFinderSettingsFlags DutyFinderSettings { get; }
